+++ WebDerm/src/test/testfunctions.h
@@ -17,7 +17,7 @@ void drawOutlineOnMouse(Mat &img);
 void Mat2Image(deque< deque<double> > &input, Mat &output);
 Mat test_normalizeColors(Mat src);
 double test_inverseGamma(int val);
-double test_Gamma(double val);
-double test_gray(int r, int g, int b);
+int test_Gamma(double val);
+int test_gray(int r, int g, int b);
 
 #endif /* TESTFUNCTIONS_H_ */
